creating ValueStateTTL and ListStateTTL Suites
