+++ sql/core/src/test/scala/org/apache/spark/sql/streaming/TransformWithStateTTLSuiteBase.scala
@@ -20,9 +20,7 @@ package org.apache.spark.sql.streaming
 import java.sql.Timestamp
 import java.time.Duration
 
-import org.apache.spark.internal.Logging
-import org.apache.spark.sql.Encoders
-import org.apache.spark.sql.execution.streaming.{ListStateImplWithTTL, MemoryStream, ValueStateImplWithTTL}
+import org.apache.spark.sql.execution.streaming.MemoryStream
 import org.apache.spark.sql.execution.streaming.state.RocksDBStateStoreProvider
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.streaming.util.StreamManualClock
@@ -40,211 +38,6 @@ case class OutputEvent(
     value: Int,
     isTTLValue: Boolean,
     ttlValue: Long)
-
-object TTLInputProcessFunction {
-  def processRow(
-      ttlMode: TTLMode,
-      row: InputEvent,
-      valueState: ValueStateImplWithTTL[Int]): Iterator[OutputEvent] = {
-    var results = List[OutputEvent]()
-    val key = row.key
-    if (row.action == "get") {
-      val currState = valueState.getOption()
-      if (currState.isDefined) {
-        results = OutputEvent(key, currState.get, isTTLValue = false, -1) :: results
-      }
-    } else if (row.action == "get_without_enforcing_ttl") {
-      val currState = valueState.getWithoutEnforcingTTL()
-      if (currState.isDefined) {
-        results = OutputEvent(key, currState.get, isTTLValue = false, -1) :: results
-      }
-    } else if (row.action == "get_ttl_value_from_state") {
-      val ttlExpiration = valueState.getTTLValue()
-      if (ttlExpiration.isDefined) {
-        results = OutputEvent(key, -1, isTTLValue = true, ttlExpiration.get) :: results
-      }
-    } else if (row.action == "put") {
-      if (ttlMode == TTLMode.EventTimeTTL() && row.eventTimeTtl != null) {
-        valueState.update(row.value, row.eventTimeTtl.getTime)
-      } else if (ttlMode == TTLMode.EventTimeTTL()) {
-        valueState.update(row.value)
-      } else {
-        valueState.update(row.value, row.ttl)
-      }
-    } else if (row.action == "get_values_in_ttl_state") {
-      val ttlValues = valueState.getValuesInTTLState()
-      ttlValues.foreach { v =>
-        results = OutputEvent(key, -1, isTTLValue = true, ttlValue = v) :: results
-      }
-    }
-
-    results.iterator
-  }
-
-  def processRow(
-    ttlMode: TTLMode,
-    row: InputEvent,
-    listState: ListStateImplWithTTL[Int]): Iterator[OutputEvent] = {
-    var results = List[OutputEvent]()
-    val key = row.key
-    if (row.action == "get") {
-      val currState = listState.get()
-      currState.foreach { v =>
-        results = OutputEvent(key, v, isTTLValue = false, -1) :: results
-      }
-    } else if (row.action == "get_without_enforcing_ttl") {
-      val currState = listState.getWithoutEnforcingTTL()
-      currState.foreach { v =>
-        results = OutputEvent(key, v, isTTLValue = false, -1) :: results
-      }
-    } else if (row.action == "get_ttl_value_from_state") {
-      val ttlExpirations = listState.getTTLValues()
-      // get the ttlExpirations that are defined and not equal to -1
-      ttlExpirations.filter(_.isDefined).foreach { ttlExpiration =>
-        results = OutputEvent(key, -1, isTTLValue = true, ttlExpiration.get) :: results
-      }
-    } else if (row.action == "put") {
-      val rowArr = Array(row.value)
-      if (ttlMode == TTLMode.EventTimeTTL() && row.eventTimeTtl != null) {
-        listState.put(rowArr, row.eventTimeTtl.getTime)
-      } else if (ttlMode == TTLMode.EventTimeTTL()) {
-        listState.put(rowArr)
-      } else {
-        listState.put(rowArr, row.ttl)
-      }
-    } else if (row.action == "append") {
-      if (ttlMode == TTLMode.EventTimeTTL() && row.eventTimeTtl != null) {
-        listState.appendValue(row.value, row.eventTimeTtl.getTime)
-      } else if (ttlMode == TTLMode.EventTimeTTL()) {
-        listState.appendValue(row.value)
-      } else {
-        listState.appendValue(row.value, row.ttl)
-      }
-    } else if (row.action == "get_values_in_ttl_state") {
-      val ttlValues = listState.getValuesInTTLState()
-      ttlValues.foreach { v =>
-        results = OutputEvent(key, -1, isTTLValue = true, ttlValue = v) :: results
-      }
-    }
-
-    results.iterator
-  }
-}
-
-class ValueStateTTLProcessor
-  extends StatefulProcessor[String, InputEvent, OutputEvent]
-  with Logging {
-
-  @transient private var _valueState: ValueStateImplWithTTL[Int] = _
-  @transient private var _ttlMode: TTLMode = _
-
-  override def init(
-      outputMode: OutputMode,
-      timeoutMode: TimeoutMode,
-      ttlMode: TTLMode): Unit = {
-    _valueState = getHandle
-      .getValueState("valueState", Encoders.scalaInt)
-      .asInstanceOf[ValueStateImplWithTTL[Int]]
-    _ttlMode = ttlMode
-  }
-
-  override def handleInputRows(
-      key: String,
-      inputRows: Iterator[InputEvent],
-      timerValues: TimerValues,
-      expiredTimerInfo: ExpiredTimerInfo): Iterator[OutputEvent] = {
-    var results = List[OutputEvent]()
-
-    for (row <- inputRows) {
-      val resultIter = TTLInputProcessFunction.processRow(_ttlMode, row, _valueState)
-      resultIter.foreach { r =>
-        results = r :: results
-      }
-    }
-
-    results.iterator
-  }
-}
-
-class ListStateTTLProcessor
-  extends StatefulProcessor[String, InputEvent, OutputEvent]
-    with Logging {
-
-  @transient private var _listState: ListStateImplWithTTL[Int] = _
-  @transient private var _ttlMode: TTLMode = _
-
-  override def init(
-   outputMode: OutputMode,
-   timeoutMode: TimeoutMode,
-   ttlMode: TTLMode): Unit = {
-    _listState = getHandle
-      .getListState("listState", Encoders.scalaInt)
-      .asInstanceOf[ListStateImplWithTTL[Int]]
-    _ttlMode = ttlMode
-  }
-
-  override def handleInputRows(
-    key: String,
-    inputRows: Iterator[InputEvent],
-    timerValues: TimerValues,
-    expiredTimerInfo: ExpiredTimerInfo): Iterator[OutputEvent] = {
-    var results = List[OutputEvent]()
-
-    for (row <- inputRows) {
-      val resultIter = TTLInputProcessFunction.processRow(_ttlMode, row, _listState)
-      resultIter.foreach { r =>
-        results = r :: results
-      }
-    }
-
-    results.iterator
-  }
-}
-
-case class MultipleValueStatesTTLProcessor(
-    ttlKey: String,
-    noTtlKey: String)
-  extends StatefulProcessor[String, InputEvent, OutputEvent]
-    with Logging {
-
-  @transient private var _valueStateWithTTL: ValueStateImplWithTTL[Int] = _
-  @transient private var _valueStateWithoutTTL: ValueStateImplWithTTL[Int] = _
-  @transient private var _ttlMode: TTLMode = _
-
-  override def init(
-      outputMode: OutputMode,
-      timeoutMode: TimeoutMode,
-      ttlMode: TTLMode): Unit = {
-    _valueStateWithTTL = getHandle
-      .getValueState("valueState", Encoders.scalaInt)
-      .asInstanceOf[ValueStateImplWithTTL[Int]]
-    _valueStateWithoutTTL = getHandle
-      .getValueState("valueState", Encoders.scalaInt)
-      .asInstanceOf[ValueStateImplWithTTL[Int]]
-    _ttlMode = ttlMode
-  }
-
-  override def handleInputRows(
-      key: String,
-      inputRows: Iterator[InputEvent],
-      timerValues: TimerValues,
-      expiredTimerInfo: ExpiredTimerInfo): Iterator[OutputEvent] = {
-    var results = List[OutputEvent]()
-    val state = if (key == ttlKey) {
-      _valueStateWithTTL
-    } else {
-      _valueStateWithoutTTL
-    }
-
-    for (row <- inputRows) {
-      val resultIterator = TTLInputProcessFunction.processRow(_ttlMode, row, state)
-      resultIterator.foreach { r =>
-        results = r :: results
-      }
-    }
-    results.iterator
-  }
-}
 
 abstract class TransformWithStateTTLSuiteBase
   extends StreamTest {
@@ -411,61 +204,6 @@ abstract class TransformWithStateTTLSuiteBase
         CheckNewAnswer(OutputEvent("k1", 1, isTTLValue = false, -1)),
         // validate that the older expiration value is removed from ttl state
         AddData(inputStream, InputEvent("k1", "get_values_in_ttl_state", -1, null)),
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer()
-      )
-    }
-  }
-
-  test("validate multiple value states - processing time ttl") {
-    withSQLConf(SQLConf.STATE_STORE_PROVIDER_CLASS.key ->
-      classOf[RocksDBStateStoreProvider].getName) {
-      val ttlKey = "k1"
-      val noTtlKey = "k2"
-
-      val inputStream = MemoryStream[InputEvent]
-      val result = inputStream.toDS()
-        .groupByKey(x => x.key)
-        .transformWithState(
-          MultipleValueStatesTTLProcessor(ttlKey, noTtlKey),
-          TimeoutMode.NoTimeouts(),
-          TTLMode.ProcessingTimeTTL())
-
-      val clock = new StreamManualClock
-      testStream(result)(
-        StartStream(Trigger.ProcessingTime("1 second"), triggerClock = clock),
-        AddData(inputStream, InputEvent(ttlKey, "put", 1, Duration.ofMinutes(1))),
-        AddData(inputStream, InputEvent(noTtlKey, "put", 2, Duration.ZERO)),
-        // advance clock to trigger processing
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(),
-        // get both state values, and make sure we get unexpired value
-        AddData(inputStream, InputEvent(ttlKey, "get", -1, null)),
-        AddData(inputStream, InputEvent(noTtlKey, "get", -1, null)),
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(
-          OutputEvent(ttlKey, 1, isTTLValue = false, -1),
-          OutputEvent(noTtlKey, 2, isTTLValue = false, -1)
-        ),
-        // ensure ttl values were added correctly, and noTtlKey has no ttl values
-        AddData(inputStream, InputEvent(ttlKey, "get_ttl_value_from_state", -1, null)),
-        AddData(inputStream, InputEvent(noTtlKey, "get_ttl_value_from_state", -1, null)),
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(OutputEvent(ttlKey, -1, isTTLValue = true, 61000)),
-        AddData(inputStream, InputEvent(ttlKey, "get_values_in_ttl_state", -1, null)),
-        AddData(inputStream, InputEvent(noTtlKey, "get_values_in_ttl_state", -1, null)),
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(OutputEvent(ttlKey, -1, isTTLValue = true, 61000)),
-        // advance clock after expiry
-        AdvanceManualClock(60 * 1000),
-        AddData(inputStream, InputEvent(ttlKey, "get", -1, null)),
-        AddData(inputStream, InputEvent(noTtlKey, "get", -1, null)),
-        // advance clock to trigger processing
-        AdvanceManualClock(1 * 1000),
-        // validate ttlKey is expired, bot noTtlKey is still present
-        CheckNewAnswer(OutputEvent(noTtlKey, 2, isTTLValue = false, -1)),
-        // validate ttl value is removed in the value state column family
-        AddData(inputStream, InputEvent(ttlKey, "get_ttl_value_from_state", -1, null)),
         AdvanceManualClock(1 * 1000),
         CheckNewAnswer()
       )
@@ -659,200 +397,6 @@ abstract class TransformWithStateTTLSuiteBase
         CheckNewAnswer(
           OutputEvent("k2", -1, isTTLValue = true, 121000),
           OutputEvent("k2", -1, isTTLValue = true, 121000))
-      )
-    }
-  }
-}
-
-class ValueStateTTLSuite extends TransformWithStateTTLSuiteBase {
-
-  import testImplicits._
-
-  override def getProcessor(): StatefulProcessor[String, InputEvent, OutputEvent] = {
-    new ValueStateTTLProcessor()
-  }
-
-  test("validate multiple value states - processing time ttl") {
-    withSQLConf(SQLConf.STATE_STORE_PROVIDER_CLASS.key ->
-      classOf[RocksDBStateStoreProvider].getName,
-      SQLConf.SHUFFLE_PARTITIONS.key -> "1") {
-      val ttlKey = "k1"
-      val noTtlKey = "k2"
-
-      val inputStream = MemoryStream[InputEvent]
-      val result = inputStream.toDS()
-        .groupByKey(x => x.key)
-        .transformWithState(
-          MultipleValueStatesTTLProcessor(ttlKey, noTtlKey),
-          TimeoutMode.NoTimeouts(),
-          TTLMode.ProcessingTimeTTL())
-
-      val clock = new StreamManualClock
-      testStream(result)(
-        StartStream(Trigger.ProcessingTime("1 second"), triggerClock = clock),
-        AddData(inputStream, InputEvent(ttlKey, "put", 1, Duration.ofMinutes(1))),
-        AddData(inputStream, InputEvent(noTtlKey, "put", 2, Duration.ZERO)),
-        // advance clock to trigger processing
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(),
-        // get both state values, and make sure we get unexpired value
-        AddData(inputStream, InputEvent(ttlKey, "get", -1, null)),
-        AddData(inputStream, InputEvent(noTtlKey, "get", -1, null)),
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(
-          OutputEvent(ttlKey, 1, isTTLValue = false, -1),
-          OutputEvent(noTtlKey, 2, isTTLValue = false, -1)
-        ),
-        // ensure ttl values were added correctly, and noTtlKey has no ttl values
-        AddData(inputStream, InputEvent(ttlKey, "get_ttl_value_from_state", -1, null)),
-        AddData(inputStream, InputEvent(noTtlKey, "get_ttl_value_from_state", -1, null)),
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(OutputEvent(ttlKey, -1, isTTLValue = true, 61000)),
-        AddData(inputStream, InputEvent(ttlKey, "get_values_in_ttl_state", -1, null)),
-        AddData(inputStream, InputEvent(noTtlKey, "get_values_in_ttl_state", -1, null)),
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(OutputEvent(ttlKey, -1, isTTLValue = true, 61000)),
-        // advance clock after expiry
-        AdvanceManualClock(60 * 1000),
-        AddData(inputStream, InputEvent(ttlKey, "get", -1, null)),
-        AddData(inputStream, InputEvent(noTtlKey, "get", -1, null)),
-        // advance clock to trigger processing
-        AdvanceManualClock(1 * 1000),
-        // validate ttlKey is expired, bot noTtlKey is still present
-        CheckNewAnswer(OutputEvent(noTtlKey, 2, isTTLValue = false, -1)),
-        // validate ttl value is removed in the value state column family
-        AddData(inputStream, InputEvent(ttlKey, "get_ttl_value_from_state", -1, null)),
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer()
-      )
-    }
-  }
-}
-
-class ListStateTTLSuite extends TransformWithStateTTLSuiteBase {
-
-  import testImplicits._
-  override def getProcessor(): StatefulProcessor[String, InputEvent, OutputEvent] = {
-    new ListStateTTLProcessor()
-  }
-
-  test("verify iterator works with expired values in middle of list - processing time ttl") {
-    withSQLConf(SQLConf.STATE_STORE_PROVIDER_CLASS.key ->
-      classOf[RocksDBStateStoreProvider].getName,
-      SQLConf.SHUFFLE_PARTITIONS.key -> "1") {
-      val inputStream = MemoryStream[InputEvent]
-      val result = inputStream.toDS()
-        .groupByKey(x => x.key)
-        .transformWithState(
-          getProcessor(),
-          TimeoutMode.NoTimeouts(),
-          TTLMode.ProcessingTimeTTL())
-
-      val clock = new StreamManualClock
-      testStream(result)(
-        StartStream(Trigger.ProcessingTime("1 second"), triggerClock = clock),
-        // Add three elements with duration of a minute
-        AddData(inputStream, InputEvent("k1", "put", 1, Duration.ofMinutes(1))),
-        AdvanceManualClock(1 * 1000),
-        AddData(inputStream, InputEvent("k1", "append", 2, Duration.ofMinutes(1))),
-        AddData(inputStream, InputEvent("k1", "append", 3, Duration.ofMinutes(1))),
-        // advance clock to trigger processing
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(),
-        // Add three elements with a duration of 15 seconds
-        AddData(inputStream, InputEvent("k1", "append", 4, Duration.ofSeconds(15))),
-        AddData(inputStream, InputEvent("k1", "append", 5, Duration.ofSeconds(15))),
-        AddData(inputStream, InputEvent("k1", "append", 6, Duration.ofSeconds(15))),
-        // advance clock to trigger processing
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(),
-        // Add three elements with a duration of a minute
-        AddData(inputStream, InputEvent("k1", "append", 7, Duration.ofMinutes(1))),
-        AddData(inputStream, InputEvent("k1", "append", 8, Duration.ofMinutes(1))),
-        AddData(inputStream, InputEvent("k1", "append", 9, Duration.ofMinutes(1))),
-        // advance clock to trigger processing
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(),
-        // Advance clock to expire the middle three elements
-        AdvanceManualClock(30 * 1000),
-        // Get all elements in the list
-        AddData(inputStream, InputEvent("k1", "get", -1, null)),
-        AdvanceManualClock(1 * 1000),
-        // Validate that the expired elements are not returned
-        CheckNewAnswer(
-          OutputEvent("k1", 1, isTTLValue = false, -1),
-          OutputEvent("k1", 2, isTTLValue = false, -1),
-          OutputEvent("k1", 3, isTTLValue = false, -1),
-          OutputEvent("k1", 7, isTTLValue = false, -1),
-          OutputEvent("k1", 8, isTTLValue = false, -1),
-          OutputEvent("k1", 9, isTTLValue = false, -1)
-        )
-      )
-    }
-  }
-
-  test("verify iterator works with expired values in middle of list - event time ttl") {
-    withSQLConf(SQLConf.STATE_STORE_PROVIDER_CLASS.key ->
-      classOf[RocksDBStateStoreProvider].getName,
-      SQLConf.SHUFFLE_PARTITIONS.key -> "1") {
-      val inputStream = MemoryStream[InputEvent]
-      val result = inputStream.toDS()
-        .withWatermark("eventTime", "1 second")
-        .groupByKey(x => x.key)
-        .transformWithState(
-          getProcessor(),
-          TimeoutMode.NoTimeouts(),
-          TTLMode.EventTimeTTL())
-
-      val eventTime1 = Timestamp.valueOf("2024-01-01 00:00:00")
-      val eventTime2 = Timestamp.valueOf("2024-01-01 00:01:00")
-      val eventTime3 = Timestamp.valueOf("2024-01-01 00:03:00")
-
-      val ttlExpirationEarly = Timestamp.valueOf("2024-01-01 00:02:00")
-      val ttlExpirationEarlyMs = ttlExpirationEarly.getTime
-
-      val ttlExpirationLate = Timestamp.valueOf("2024-01-01 00:05:00")
-      val ttlExpirationLateMs = ttlExpirationLate.getTime
-
-      val clock = new StreamManualClock
-      testStream(result)(
-        StartStream(Trigger.ProcessingTime("1 second"), triggerClock = clock),
-        // Add three elements with duration of a minute
-        AddData(inputStream, InputEvent("k1", "put", 1, null, eventTime1, ttlExpirationLate)),
-        AdvanceManualClock(1 * 1000),
-        AddData(inputStream, InputEvent("k1", "append", 2, null, eventTime1, ttlExpirationLate)),
-        AddData(inputStream, InputEvent("k1", "append", 3, null, eventTime1, ttlExpirationLate)),
-        // advance clock to trigger processing
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(),
-        // Add three elements with a duration of 15 seconds
-        AddData(inputStream, InputEvent("k1", "append", 4, null, eventTime2, ttlExpirationEarly)),
-        AddData(inputStream, InputEvent("k1", "append", 5, null, eventTime2, ttlExpirationEarly)),
-        AddData(inputStream, InputEvent("k1", "append", 6, null, eventTime2, ttlExpirationEarly)),
-        // advance clock to trigger processing
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(),
-        // Add three elements with a duration of a minute
-        AddData(inputStream, InputEvent("k1", "append", 7, null, eventTime3, ttlExpirationLate)),
-        AddData(inputStream, InputEvent("k1", "append", 8, null, eventTime3, ttlExpirationLate)),
-        AddData(inputStream, InputEvent("k1", "append", 9, null, eventTime3, ttlExpirationLate)),
-        // advance clock to trigger processing
-        AdvanceManualClock(1 * 1000),
-        CheckNewAnswer(),
-        // Advance clock to expire the middle three elements
-        AdvanceManualClock(30 * 1000),
-        // Get all elements in the list
-        AddData(inputStream, InputEvent("k1", "get", -1, null)),
-        AdvanceManualClock(1 * 1000),
-        // Validate that the expired elements are not returned
-        CheckNewAnswer(
-          OutputEvent("k1", 1, isTTLValue = false, -1),
-          OutputEvent("k1", 2, isTTLValue = false, -1),
-          OutputEvent("k1", 3, isTTLValue = false, -1),
-          OutputEvent("k1", 7, isTTLValue = false, -1),
-          OutputEvent("k1", 8, isTTLValue = false, -1),
-          OutputEvent("k1", 9, isTTLValue = false, -1)
-        )
       )
     }
   }
